requirements.txt
confluent-kafka==2.6.1
docker==7.1.0
Jinja2==3.1.4
python-dotenv==1.0.1